refactor: moved logic for toggling widgets to widgetslayout component
