+++ layout/WidgetsLayout.js
@@ -1,201 +1,67 @@
-/*
-  This example requires some changes to your config:
-  
-  ```
-  // tailwind.config.js
-  module.exports = {
-    // ...
-    plugins: [
-      // ...
-      require('@tailwindcss/aspect-ratio'),
-    ],
-  }
-  ```
-*/
-"use client";
-
-import { useState } from "react";
-import { Dialog, DialogPanel, TransitionChild } from "@headlessui/react";
-import { HeartIcon, XMarkIcon } from "@heroicons/react/24/outline";
-import { PencilIcon, PlusIcon } from "@heroicons/react/20/solid";
+import { observer } from "mobx-react-lite";
+import { Show } from "@/components/ui/Show";
 import { Container } from "@/components/ui/Container";
+import { showUp, toggleWidget } from "@/library/animations";
+import { useGeneralStore } from "@/providers/generalStore";
 
+const BackDrop = (props) => (
+  <Container.Animated
+    className="absolute top-0 h-screen w-screen bg-backDrop"
+    {...showUp}
+    {...props}
+  />
+);
+
+const WidgetsToggle = ({ isWidgetsOn, setIsWidgetsOn }) => (
+  <Container.Animated
+    onMouseEnter={() => setIsWidgetsOn(!isWidgetsOn)}
+    onMouseLeave={() => setIsWidgetsOn(!isWidgetsOn)}
+    {...toggleWidget(isWidgetsOn)}
+    // component={isWidgetsOn && <WidgetsLayout />}
+  />
+);
 const WidgetsLayout = () => {
-  const [open, setOpen] = useState(true);
+  const { isWidgetsOn, setIsWidgetsOn } = useGeneralStore();
 
   return (
-    <Dialog open={open} onClose={setOpen} className="relative z-10">
-      <Container className="absolute z-50 h-full w-full border bg-black" />
-      <div className="fixed inset-0 overflow-hidden">
-        <div className="absolute inset-0 overflow-hidden">
-          <div className="pointer-events-none fixed inset-y-0 right-0 flex max-w-full pl-10">
-            <DialogPanel
-              transition
-              className="pointer-events-auto relative w-96 transform transition duration-500 ease-in-out data-[closed]:translate-x-full sm:duration-700"
-            >
-              <TransitionChild>
-                <div className="absolute left-0 top-0 -ml-8 flex pr-2 pt-4 duration-500 ease-in-out data-[closed]:opacity-0 sm:-ml-10 sm:pr-4">
-                  <button
-                    type="button"
-                    onClick={() => setOpen(false)}
-                    className="relative rounded-md text-gray-300 hover:text-white focus:outline-none focus:ring-2 focus:ring-white"
-                  >
-                    <span className="absolute -inset-2.5" />
-                    <span className="sr-only">Close panel</span>
-                    <XMarkIcon aria-hidden="true" className="h-6 w-6" />
-                  </button>
-                </div>
-              </TransitionChild>
-              <div className="h-full overflow-y-auto bg-white p-8">
-                <div className="space-y-6 pb-16">
-                  <div>
-                    <div className="aspect-h-7 aspect-w-10 block w-full overflow-hidden rounded-lg">
-                      <img
-                        alt=""
-                        src="https://images.unsplash.com/photo-1582053433976-25c00369fc93?ixid=MXwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHw%3D&ixlib=rb-1.2.1&auto=format&fit=crop&w=512&q=80"
-                        className="object-cover"
-                      />
-                    </div>
-                    <div className="mt-4 flex items-start justify-between">
-                      <div>
-                        <h2 className="text-base font-semibold leading-6 text-gray-900">
-                          <span className="sr-only">Details for </span>
-                          IMG_4985.HEIC
-                        </h2>
-                        <p className="text-sm font-medium text-gray-500">
-                          3.9 MB
-                        </p>
-                      </div>
-                      <button
-                        type="button"
-                        className="relative ml-4 flex h-8 w-8 items-center justify-center rounded-full bg-white text-gray-400 hover:bg-gray-100 hover:text-gray-500 focus:outline-none focus:ring-2 focus:ring-indigo-500"
-                      >
-                        <span className="absolute -inset-1.5" />
-                        <HeartIcon aria-hidden="true" className="h-6 w-6" />
-                        <span className="sr-only">Favorite</span>
-                      </button>
-                    </div>
-                  </div>
-                  <div>
-                    <h3 className="font-medium text-gray-900">Information</h3>
-                    <dl className="mt-2 divide-y divide-gray-200 border-b border-t border-gray-200">
-                      <div className="flex justify-between py-3 text-sm font-medium">
-                        <dt className="text-gray-500">Uploaded by</dt>
-                        <dd className="text-gray-900">Marie Culver</dd>
-                      </div>
-                      <div className="flex justify-between py-3 text-sm font-medium">
-                        <dt className="text-gray-500">Created</dt>
-                        <dd className="text-gray-900">June 8, 2020</dd>
-                      </div>
-                      <div className="flex justify-between py-3 text-sm font-medium">
-                        <dt className="text-gray-500">Last modified</dt>
-                        <dd className="text-gray-900">June 8, 2020</dd>
-                      </div>
-                      <div className="flex justify-between py-3 text-sm font-medium">
-                        <dt className="text-gray-500">Dimensions</dt>
-                        <dd className="text-gray-900">4032 x 3024</dd>
-                      </div>
-                      <div className="flex justify-between py-3 text-sm font-medium">
-                        <dt className="text-gray-500">Resolution</dt>
-                        <dd className="text-gray-900">72 x 72</dd>
-                      </div>
-                    </dl>
-                  </div>
-                  <div>
-                    <h3 className="font-medium text-gray-900">Description</h3>
-                    <div className="mt-2 flex items-center justify-between">
-                      <p className="text-sm italic text-gray-500">
-                        Add a description to this image.
-                      </p>
-                      <button
-                        type="button"
-                        className="relative -mr-2 flex h-8 w-8 items-center justify-center rounded-full bg-white text-gray-400 hover:bg-gray-100 hover:text-gray-500 focus:outline-none focus:ring-2 focus:ring-indigo-500"
-                      >
-                        <span className="absolute -inset-1.5" />
-                        <PencilIcon aria-hidden="true" className="h-5 w-5" />
-                        <span className="sr-only">Add description</span>
-                      </button>
-                    </div>
-                  </div>
-                  <div>
-                    <h3 className="font-medium text-gray-900">Shared with</h3>
-                    <ul
-                      role="list"
-                      className="mt-2 divide-y divide-gray-200 border-b border-t border-gray-200"
-                    >
-                      <li className="flex items-center justify-between py-3">
-                        <div className="flex items-center">
-                          <img
-                            alt=""
-                            src="https://images.unsplash.com/photo-1502685104226-ee32379fefbe?ixlib=rb-=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=facearea&facepad=3&w=1024&h=1024&q=80"
-                            className="h-8 w-8 rounded-full"
-                          />
-                          <p className="ml-4 text-sm font-medium text-gray-900">
-                            Aimee Douglas
-                          </p>
-                        </div>
-                        <button
-                          type="button"
-                          className="ml-6 rounded-md bg-white text-sm font-medium text-indigo-600 hover:text-indigo-500 focus:outline-none focus:ring-2 focus:ring-indigo-500 focus:ring-offset-2"
-                        >
-                          Remove<span className="sr-only"> Aimee Douglas</span>
-                        </button>
-                      </li>
-                      <li className="flex items-center justify-between py-3">
-                        <div className="flex items-center">
-                          <img
-                            alt=""
-                            src="https://images.unsplash.com/photo-1494790108377-be9c29b29330?ixlib=rb-1.2.1&ixqx=oilqXxSqey&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=facearea&facepad=2&w=256&h=256&q=80"
-                            className="h-8 w-8 rounded-full"
-                          />
-                          <p className="ml-4 text-sm font-medium text-gray-900">
-                            Andrea McMillan
-                          </p>
-                        </div>
-                        <button
-                          type="button"
-                          className="ml-6 rounded-md bg-white text-sm font-medium text-indigo-600 hover:text-indigo-500 focus:outline-none focus:ring-2 focus:ring-indigo-500 focus:ring-offset-2"
-                        >
-                          Remove
-                          <span className="sr-only"> Andrea McMillan</span>
-                        </button>
-                      </li>
-                      <li className="flex items-center justify-between py-2">
-                        <button
-                          type="button"
-                          className="group -ml-1 flex items-center rounded-md bg-white p-1 focus:outline-none focus:ring-2 focus:ring-indigo-500"
-                        >
-                          <span className="flex h-8 w-8 items-center justify-center rounded-full border-2 border-dashed border-gray-300 text-gray-400">
-                            <PlusIcon aria-hidden="true" className="h-5 w-5" />
-                          </span>
-                          <span className="ml-4 text-sm font-medium text-indigo-600 group-hover:text-indigo-500">
-                            Share
-                          </span>
-                        </button>
-                      </li>
-                    </ul>
-                  </div>
-                  <div className="flex">
-                    <button
-                      type="button"
-                      className="flex-1 rounded-md bg-indigo-600 px-3 py-2 text-sm font-semibold text-white shadow-sm hover:bg-indigo-500 focus-visible:outline focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-indigo-600"
-                    >
-                      Download
-                    </button>
-                    <button
-                      type="button"
-                      className="ml-3 flex-1 rounded-md bg-white px-3 py-2 text-sm font-semibold text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 hover:bg-gray-50"
-                    >
-                      Delete
-                    </button>
-                  </div>
-                </div>
+    <>
+      <WidgetsToggle
+        isWidgetsOn={isWidgetsOn}
+        setIsWidgetsOn={setIsWidgetsOn}
+      />
+      <Show isTrue={isWidgetsOn} animatePresence>
+        <BackDrop />
+      </Show>
+      {/* <Container className="scrollbar-hide fixed inset-y-0 right-0 flex h-full w-96 overflow-auto bg-white py-16">
+        <Container className="mx-auto flex w-4/5 flex-col space-y-6 border">
+          <div>
+            <div className="aspect-h-7 aspect-w-6 block w-full overflow-hidden rounded-lg">
+              <img
+                alt=""
+                src="https://images.unsplash.com/photo-1582053433976-25c00369fc93?ixid=MXwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHw%3D&ixlib=rb-1.2.1&auto=format&fit=crop&w=512&q=80"
+                className="object-cover"
+              />
+            </div>
+            <div className="mt-4 flex items-start justify-between">
+              <div>
+                <h2 className="text-base font-semibold leading-6 text-gray-900">
+                  <span className="sr-only">Details for </span>
+                  IMG_4985.HEIC
+                </h2>
+                <p className="text-sm font-medium text-gray-500">3.9 MB</p>
               </div>
-            </DialogPanel>
+              <button
+                type="button"
+                className="relative ml-4 flex h-8 w-8 items-center justify-center rounded-full bg-white text-gray-400 hover:bg-gray-100 hover:text-gray-500 focus:outline-none focus:ring-2 focus:ring-indigo-500"
+              >
+                <span className="absolute -inset-1.5" />
+                <span className="sr-only">Favorite</span>
+              </button>
+            </div>
           </div>
-        </div>
-      </div>
-    </Dialog>
+        </Container>
+      </Container> */}
+    </>
   );
 };
-export default WidgetsLayout;
+export default observer(WidgetsLayout);
